animedia/backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
Flask
request
flask-cors
# For database migrations